reset checkpoint
