--- environment/OPERATIONS_PHASE1_SUMMARY.md
@@ -1,204 +1,0 @@
-# Operations Framework - Phase 1 Implementation Summary
-
-## 🎯 Overview
-
-Successfully implemented Phase 1 of the Operations-based workflow execution system as outlined in the PRD. This establishes the foundation for deterministic, replayable workflow execution in Dive.
-
-## ✅ Phase 1 Deliverables Completed
-
-### 1. Core Operation Types and Interfaces
-
-**Files Created:**
-- `environment/operation.go` - Core operation types and interfaces
-
-**Key Components:**
-- `Operation` struct with deterministic ID generation
-- `OperationResult` for caching execution results
-- `OperationExecutor` interface for execution with recording/replay
-- `NewOperation()` helper function with deterministic hash-based IDs
-
-```go
-type Operation struct {
-    ID         OperationID            // Deterministic, unique identifier
-    Type       string                 // "agent_response", "action", etc.
-    StepName   string                 // Workflow step that triggered this
-    PathID     string                 // Execution path identifier  
-    Parameters map[string]interface{} // Input parameters
-}
-```
-
-### 2. Event System Extension
-
-**Files Modified:**
-- `environment/execution_event.go` - Added new event types
-
-**New Event Types Added:**
-- `EventOperationStarted` - Operation begins execution
-- `EventOperationCompleted` - Operation completes successfully
-- `EventOperationFailed` - Operation fails with error
-- `EventStateSet` - State mutation events
-- `EventStateDeleted` - State deletion events
-- `EventTimeAccessed` - Deterministic time access
-- `EventRandomGenerated` - Random value generation
-
-### 3. State Management System
-
-**Files Created:**
-- `environment/workflow_state.go` - Controlled state management
-
-**Key Components:**
-- `WorkflowState` struct with controlled mutations
-- `RisorStateObject` for script integration
-- Event recording for all state changes
-- Thread-safe operations with mutex protection
-
-```go
-type WorkflowState struct {
-    executionID string
-    recorder    ExecutionRecorder
-    values      map[string]interface{}
-    mutex       sync.RWMutex
-}
-```
-
-### 4. EventBasedExecution Integration
-
-**Files Modified:**
-- `environment/execution.go` - Extended with operation support
-
-**Key Enhancements:**
-- Implemented `OperationExecutor` interface
-- Added operation result caching
-- Integrated `WorkflowState` with Risor script globals
-- Updated `handlePromptStep()` and `handleActionStep()` to use operations
-- Added current path ID tracking for operations
-
-### 5. Comprehensive Testing
-
-**Files Created:**
-- `environment/operation_test.go` - Complete test suite
-
-**Test Coverage:**
-- Operation execution and caching
-- Error handling and recording
-- State management operations
-- Deterministic ID generation
-- Risor state object integration
-
-## 🔧 Key Features Implemented
-
-### Deterministic Operation Execution
-
-Operations clearly separate deterministic preparation from non-deterministic execution:
-
-```go
-// Deterministic: prepare parameters
-params := evaluateParams(step.Parameters())
-
-// Operation: execute action (non-deterministic)  
-op := NewOperation("action_execution", step.Name(), pathID, params)
-result, err := execution.ExecuteOperation(ctx, op, func() (interface{}, error) {
-    return action.Execute(ctx, params)
-})
-
-// Deterministic: process result
-return processResult(result)
-```
-
-### Automatic Event Recording
-
-All operations are automatically recorded with:
-- Operation start/complete/failure events
-- Deterministic operation IDs for replay
-- Duration tracking and error capture
-- Parameter and result serialization
-
-### State Management with Event Tracking
-
-State mutations are controlled and recorded:
-
-```go
-// In Risor scripts:
-state.set("processed_items", 42)       // Recorded as EventStateSet
-state.delete("temporary_data")         // Recorded as EventStateDeleted  
-value = state.get("processed_items")   // Read-only, no event
-```
-
-### Replay Support Foundation
-
-Operations can be replayed from recorded events:
-- Operation results are cached for immediate replay
-- Deterministic IDs ensure consistent operation matching
-- Replay mode skips recording duplicate events
-
-## 🚀 Usage Examples
-
-### Basic Operation Execution
-
-```go
-op := NewOperation(
-    "agent_response",           // Operation type
-    "process_input",           // Step name  
-    "path-1",                 // Path ID
-    map[string]interface{}{   // Parameters
-        "prompt": "Process this data",
-        "agent": "my-agent",
-    },
-)
-
-result, err := execution.ExecuteOperation(ctx, op, func() (interface{}, error) {
-    return agent.CreateResponse(ctx, dive.WithMessage(prompt))
-})
-```
-
-### State Management in Scripts
-
-```javascript
-// Risor script with state management
-state.set("current_step", "processing")
-state.set("items_processed", 0)
-
-for item in items {
-    // Process item...
-    count = state.get("items_processed") 
-    state.set("items_processed", count + 1)
-}
-
-state.set("current_step", "completed")
-```
-
-## 📊 Test Results
-
-All tests pass successfully:
-- ✅ Operation execution and caching
-- ✅ Error handling and recording  
-- ✅ Deterministic ID generation
-- ✅ State management operations
-- ✅ Risor integration
-- ✅ Backward compatibility with existing workflows
-
-## 🔄 Backward Compatibility
-
-Phase 1 maintains full backward compatibility:
-- Existing workflows continue to work unchanged
-- New operation recording is transparent to existing code
-- State management is additive (existing script globals unchanged)
-- Event recording is non-breaking
-
-## 🎯 Next Steps (Phase 2)
-
-Phase 1 establishes the foundation for Phase 2 objectives:
-- Refactor all non-deterministic operations (LLM calls, actions, time/random access)
-- Implement complete replay capabilities with validation
-- Add workflow versioning support
-- Create debugging and inspection tools
-
-## 📈 Success Metrics
-
-✅ **Operation framework implemented and tested**
-✅ **State management system functional**  
-✅ **Existing workflows continue to work**
-✅ **Event recording captures all operations**
-✅ **Performance impact minimal** (operations add ~microsecond overhead)
-
-The operations framework is now ready for Phase 2 migration of existing workflow execution patterns.
